Django-School-Management-System-master/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
colorama==0.4.4
Django==3.2.5
django-widget-tweaks==1.4.8
flake8==3.9.2
isort==5.9.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.3.2
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.2
tomli==1.0.4

djangorestframework~=3.12.4